refactor(WinningStatusˆ): 패키지 변경
