+++ src/test/java/domain/result/ResultTest.java
@@ -3,7 +3,6 @@ package domain.result;
 import static util.TestUtil.createAmount10000Result;
 
 import domain.Money;
-import domain.WinningStatus;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
